Move appearance sections to organizations' settings form (#14736)
