+++ decidim-admin/app/forms/decidim/admin/organization_form.rb
@@ -14,11 +14,6 @@ module Decidim
       translatable_attribute :description, Decidim::Attributes::RichText
       attribute :reference_prefix, String
       attribute :time_zone, String
-      attribute :twitter_handler, String
-      attribute :facebook_handler, String
-      attribute :instagram_handler, String
-      attribute :youtube_handler, String
-      attribute :github_handler, String
       attribute :default_locale, String
       attribute :badges_enabled, Boolean
       attribute :comments_max_length, Integer, default: 0
@@ -27,8 +22,32 @@ module Decidim
       attribute :machine_translation_display_priority, String
       attribute :enable_participatory_space_filters, Boolean
 
+      attribute :twitter_handler, String
+      attribute :facebook_handler, String
+      attribute :instagram_handler, String
+      attribute :youtube_handler, String
+      attribute :github_handler, String
+
+      attribute :logo
+      attribute :remove_logo, Boolean, default: false
+      attribute :favicon
+      attribute :remove_favicon, Boolean, default: false
+      attribute :official_url
+      attribute :official_img_footer
+      attribute :remove_official_img_footer, Boolean, default: false
+
+      attribute :primary_color, String
+      attribute :secondary_color, String
+      attribute :tertiary_color, String
+
       attribute :send_welcome_notification, Boolean
       attribute :customize_welcome_notification, Boolean
+
+      attribute :enable_omnipresent_banner, Boolean, default: false
+      attribute :omnipresent_banner_url, String
+
+      translatable_attribute :omnipresent_banner_title, String
+      translatable_attribute :omnipresent_banner_short_description, String
 
       translatable_attribute :welcome_notification_subject, String
       translatable_attribute :welcome_notification_body, Decidim::Attributes::RichText
@@ -47,6 +66,16 @@ module Decidim
       validates :machine_translation_display_priority,
                 inclusion: { in: Decidim::Organization::AVAILABLE_MACHINE_TRANSLATION_DISPLAY_PRIORITIES },
                 if: :machine_translation_enabled?
+
+      validates :official_img_footer,
+                :logo,
+                passthru: { to: Decidim::Organization }
+
+      validates :omnipresent_banner_url, url: true, presence: true, if: :enable_omnipresent_banner?
+      validates :omnipresent_banner_title, translatable_presence: true, if: :enable_omnipresent_banner?
+      validates :omnipresent_banner_short_description, translatable_presence: true, if: :enable_omnipresent_banner?
+
+      alias enable_omnipresent_banner? enable_omnipresent_banner
 
       def machine_translation_priorities
         Decidim::Organization::AVAILABLE_MACHINE_TRANSLATION_DISPLAY_PRIORITIES.map do |priority|
