+++ decidim-admin/app/packs/entrypoints/decidim_admin.js
@@ -5,7 +5,6 @@ import "src/decidim/admin/tab_focus"
 import initLanguageChangeSelect from "src/decidim/admin/choose_language"
 import "src/decidim/admin/application"
 import "src/decidim/admin/resources_permissions"
-import "src/decidim/admin/welcome_notification"
 import "src/decidim/admin/newsletters"
 import "src/decidim/admin/form"
 import "src/decidim/admin/external_domain_allowlist"
