+++ decidim-admin/app/views/decidim/admin/organization/form/_colors.html.erb
@@ -3,7 +3,7 @@
     <button class="card-divider-button" data-open="true" data-controls="panel-colors" type="button">
       <%= icon "arrow-right-s-line" %>
       <h2 class="card-title" id="colors">
-        <%= t("colors_title", scope: "decidim.admin.organization_appearance.form.colors") %>
+        <%= t("colors_title", scope: "decidim.admin.organization.form.colors") %>
       </h2>
     </button>
   </div>
@@ -11,24 +11,24 @@
     <div class="row column">
       <details class="color-picker">
         <summary>
-          <%= t(".title") %>
+          <%= t("title", scope: "decidim.admin.organization.form.colors") %>
         </summary>
 
         <div class="color-picker__box">
-          <p><%= t(".explanation") %></p>
+          <p><%= t("explanation", scope: "decidim.admin.organization.form.colors") %></p>
 
           <label class="color-picker__label_choose">
-            <%= t(".choose_color") %>
+            <%= t("choose_color", scope: "decidim.admin.organization.form.colors") %>
             <div id="primary-selector"></div>
           </label>
 
           <label class="color-picker__label_saturation" for="primary-saturation">
-            <%= t(".saturation") %>
+            <%= t("saturation", scope: "decidim.admin.organization.form.colors") %>
             <input id="primary-saturation" type="range" min="0" max="100">
           </label>
 
           <fieldset>
-            <legend><%= t(".legend_html") %></legend>
+            <legend><%= t("legend_html", scope: "decidim.admin.organization.form.colors") %></legend>
 
             <label>
               <%= t("primary_color", scope: "activemodel.attributes.organization") %>
@@ -49,12 +49,12 @@
             </label>
 
             <button id="set-colors" class="button button__sm button__secondary">
-              <%= t(".update_suggested_colors") %>
+              <%= t("update_suggested_colors", scope: "decidim.admin.organization.form.colors") %>
             </button>
           </fieldset>
         </div>
       </details>
-      <%= cell("decidim/announcement", t(".colors_warning_html"), callout_class: "warning") %>
+      <%= cell("decidim/announcement", t("colors_warning_html", scope: "decidim.admin.organization.form.colors"), callout_class: "warning") %>
 
       <div class="organization-colors">
         <%= form.color_field :primary_color, value: current_organization.colors["primary"] %>
