+++ decidim-admin/spec/commands/decidim/admin/update_organization_appearance_spec.rb
@@ -5,14 +5,13 @@ require "spec_helper"
 module Decidim::Admin
   describe UpdateOrganizationAppearance do
     describe "call" do
-      let(:organization) { create(:organization, enable_omnipresent_banner: true) }
+      let(:organization) { create(:organization, highlighted_content_banner_enabled: true) }
       let(:user) { create(:user, organization:) }
       let(:params) do
         {
           organization: {
-            enable_omnipresent_banner: false,
-            header_snippets: '<script>alert("Hello");</script>',
-            favicon: upload_test_file(Decidim::Dev.test_file("icon.png", "image/png"))
+            highlighted_content_banner_enabled: false,
+            header_snippets: '<script>alert("Hello");</script>'
           }
         }
       end
@@ -40,7 +39,7 @@ module Decidim::Admin
           command.call
           organization.reload
 
-          expect(organization.enable_omnipresent_banner).to be_truthy
+          expect(organization.highlighted_content_banner_enabled).to be_truthy
         end
       end
 
@@ -48,17 +47,10 @@ module Decidim::Admin
         before do
           allow(form).to receive(:invalid?).and_return(false)
           expect(organization).to receive(:valid?).at_least(:once).and_return(false)
-          organization.errors.add(:official_img_footer, "File resolution is too large")
         end
 
         it "broadcasts invalid" do
           expect { command.call }.to broadcast(:invalid)
-        end
-
-        it "adds errors to the form" do
-          command.call
-
-          expect(form.errors[:official_img_footer]).not_to be_empty
         end
       end
 
@@ -84,7 +76,7 @@ module Decidim::Admin
           expect { command.call }.to broadcast(:ok)
           organization.reload
 
-          expect(organization.enable_omnipresent_banner).to be_falsey
+          expect(organization.highlighted_content_banner_enabled).to be_falsey
         end
 
         it "does not save header snippets" do
@@ -104,15 +96,6 @@ module Decidim::Admin
             organization.reload
 
             expect(organization.header_snippets).to be_present
-          end
-        end
-
-        context "when there is a favicon in the params" do
-          it "does set a favicon for the organization" do
-            command.call
-            organization.reload
-
-            expect(organization.attached_uploader(:favicon).variant_url(:small)).to be_present
           end
         end
       end
