+++ decidim-admin/spec/forms/base_organization_form_spec.rb
@@ -38,6 +38,14 @@ module Decidim
       end
       let(:organization) { create(:organization) }
       let(:available_locales) { organization.available_locales }
+      let(:empty_traslatable_attribute) do
+        { en: "", es: "", ca: "" }
+      end
+      let(:enable_omnipresent_banner) { false }
+      let(:omnipresent_banner_url) { nil }
+      let(:omnipresent_banner_title) { empty_traslatable_attribute }
+      let(:omnipresent_banner_short_description) { empty_traslatable_attribute }
+
       let(:attributes) do
         {
           "organization" => {
@@ -57,7 +65,15 @@ module Decidim
             "machine_translation_display_priority" => translation_priority,
             "admin_terms_of_service_body_ca" => admin_terms_of_service_body[:ca],
             "admin_terms_of_service_body_en" => admin_terms_of_service_body[:en],
-            "admin_terms_of_service_body_es" => admin_terms_of_service_body[:es]
+            "admin_terms_of_service_body_es" => admin_terms_of_service_body[:es],
+            "enable_omnipresent_banner" => enable_omnipresent_banner,
+            "omnipresent_banner_url" => omnipresent_banner_url,
+            "omnipresent_banner_title_en" => omnipresent_banner_title[:en],
+            "omnipresent_banner_title_es" => omnipresent_banner_title[:es],
+            "omnipresent_banner_title_ca" => omnipresent_banner_title[:ca],
+            "omnipresent_banner_short_description_en" => omnipresent_banner_short_description[:en],
+            "omnipresent_banner_short_description_es" => omnipresent_banner_short_description[:es],
+            "omnipresent_banner_short_description_ca" => omnipresent_banner_short_description[:ca]
           }
         }
       end
@@ -120,6 +136,37 @@ module Decidim
         end
 
         it { is_expected.to be_invalid }
+      end
+
+      context "when enable_omnipresent_banner is true" do
+        let(:enable_omnipresent_banner) { true }
+        let(:omnipresent_banner_url) { "http://www.example.org/random_url" }
+        let(:omnipresent_banner_title) do
+          { en: "title", es: "título", ca: "títol" }
+        end
+        let(:omnipresent_banner_short_description) do
+          { en: "description", es: "descripción", ca: "descripció" }
+        end
+
+        it { is_expected.to be_valid }
+
+        context "and omnipresent_banner_url is blank" do
+          let(:omnipresent_banner_url) { nil }
+
+          it { is_expected.not_to be_valid }
+        end
+
+        context "and omnipresent_banner_title is blank" do
+          let(:omnipresent_banner_title) { empty_traslatable_attribute }
+
+          it { is_expected.not_to be_valid }
+        end
+
+        context "and omnipresent_banner_short_description is blank" do
+          let(:omnipresent_banner_short_description) { empty_traslatable_attribute }
+
+          it { is_expected.not_to be_valid }
+        end
       end
     end
   end
