+++ decidim-admin/spec/forms/organization_appearance_form_spec.rb
@@ -22,13 +22,6 @@ module Decidim
       let(:highlighted_content_banner_action_subtitle) { empty_translatable }
       let(:highlighted_content_banner_action_url) { nil }
       let(:highlighted_content_banner_image) { nil }
-      let(:enable_omnipresent_banner) { false }
-      let(:omnipresent_banner_url) { nil }
-      let(:empty_traslatable_attribute) do
-        { en: "", es: "", ca: "" }
-      end
-      let(:omnipresent_banner_title) { empty_traslatable_attribute }
-      let(:omnipresent_banner_short_description) { empty_traslatable_attribute }
       let(:attributes) do
         {
           "organization" => {
@@ -49,15 +42,7 @@ module Decidim
             "highlighted_content_banner_action_subtitle_es" => highlighted_content_banner_action_subtitle[:es],
             "highlighted_content_banner_action_subtitle_ca" => highlighted_content_banner_action_subtitle[:ca],
             "highlighted_content_banner_action_url" => highlighted_content_banner_action_url,
-            "highlighted_content_banner_image" => highlighted_content_banner_image,
-            "enable_omnipresent_banner" => enable_omnipresent_banner,
-            "omnipresent_banner_url" => omnipresent_banner_url,
-            "omnipresent_banner_title_en" => omnipresent_banner_title[:en],
-            "omnipresent_banner_title_es" => omnipresent_banner_title[:es],
-            "omnipresent_banner_title_ca" => omnipresent_banner_title[:ca],
-            "omnipresent_banner_short_description_en" => omnipresent_banner_short_description[:en],
-            "omnipresent_banner_short_description_es" => omnipresent_banner_short_description[:es],
-            "omnipresent_banner_short_description_ca" => omnipresent_banner_short_description[:ca]
+            "highlighted_content_banner_image" => highlighted_content_banner_image
           }
         }
       end
@@ -137,37 +122,6 @@ module Decidim
         let(:cta_button_path) { "processes/my_process/" }
 
         it { is_expected.to be_valid }
-      end
-
-      context "when enable_omnipresent_banner is true" do
-        let(:enable_omnipresent_banner) { true }
-        let(:omnipresent_banner_url) { "http://www.example.org/random_url" }
-        let(:omnipresent_banner_title) do
-          { en: "title", es: "título", ca: "títol" }
-        end
-        let(:omnipresent_banner_short_description) do
-          { en: "description", es: "descripción", ca: "descripció" }
-        end
-
-        it { is_expected.to be_valid }
-
-        context "and omnipresent_banner_url is blank" do
-          let(:omnipresent_banner_url) { nil }
-
-          it { is_expected.not_to be_valid }
-        end
-
-        context "and omnipresent_banner_title is blank" do
-          let(:omnipresent_banner_title) { empty_traslatable_attribute }
-
-          it { is_expected.not_to be_valid }
-        end
-
-        context "and omnipresent_banner_short_description is blank" do
-          let(:omnipresent_banner_short_description) { empty_traslatable_attribute }
-
-          it { is_expected.not_to be_valid }
-        end
       end
     end
   end
